Revert "Fix: lib: add missing BT_ASSERT_PRE_NO_ERROR in trace-ir/clock-class.c"
This reverts commit bbc5046d02a09977fb1743ad26b7fcedeb888850.

After further discussion, we concluded that the existing behavior
(before the reverted patch) is desirable.  It should only be forbidden
to call an API function while there is a current error for API functions
that can fail and set a current error themselves.

The documentation currently says:

    You cannot call any libbabeltrace2 function when the current thread
    has an error, except the reference counting functions (ending with
    _get_ref() or _put_ref()).

It will later be updated to reflect the fact that it's allowed to call
API functions that can't fail while there is a current error.

Change-Id: Id7d1419862066e0e3d9fd537198b7e630a9c2dc8
Reviewed-on: https://review.lttng.org/c/babeltrace/+/10034
Reviewed-by: Philippe Proulx <eeppeliteloop@gmail.com>
CI-Build: Simon Marchi <simon.marchi@efficios.com>
Tested-by: jenkins <jenkins@lttng.org>