build: move script to generate version.i to its own file
Move the script from src/common/Makefile.am to its own file.  This will
make it easier to read, modify and lint.

I inlined the big comment from Makefile.am in the new script.  I tried
to keep the code as-is as much as possible, but I needed to tweak a few
things.

 - Replace $(FOO) with $FOO
 - The script expects to get the top srcdir and the grep and sed
   binaries through the TOP_SRCDIR, GREP and SED env var.  GREP and SED
   are optional, to make it easier to run the script by hand

Change-Id: Iaa8948018785e0de4250285110b78ccea1abaedb
Signed-off-by: Simon Marchi <simon.marchi@efficios.com>
Reviewed-on: https://review.lttng.org/c/babeltrace/+/10095
Reviewed-by: Michael Jeanson <mjeanson@efficios.com>