doc: fix passing non-event messages in distill examples
The distill filter example means to let non-event messages go through,
but it actually drops them.  Fix it so it actually lets them through.

Verified by sending the output of the filter to `sink.text.details`.
Before the patch, there is no stream beginning message.  After the
patch, there is.

Change-Id: I2414a37ca7fc5fa6e5ed0fe179ee39cadf261f31
Signed-off-by: Simon Marchi <simon.marchi@efficios.com>
Reviewed-on: https://review.lttng.org/c/babeltrace/+/10444
Tested-by: jenkins <jenkins@lttng.org>
Reviewed-by: Philippe Proulx <eeppeliteloop@gmail.com>